Move the chat health check useEffect into the root of the App so that it is not de-rendered by DOM when chat is disabled.
